fix: 修复selectCount返回类型(Integer→Long)，注释掉SecurityUtils依赖代码(demo演示用)

+++ src/main/java/wo1261931780/stdataAuthorityControl/config/DataPermissionMapper.java
@@ -65,7 +65,7 @@ public interface DataPermissionMapper<T> extends BaseMapper<T> {
 	 */
 	@Override
 	@UserDataPermission
-	Integer selectCount(@Param(Constants.WRAPPER) Wrapper<T> queryWrapper);
+	Long selectCount(@Param(Constants.WRAPPER) Wrapper<T> queryWrapper);
 
 	/**
 	 * 根据 entity 条件，查询全部记录
